remove002: 장바구니 ê³관련 삭제

+++ src/main/java/com/mysite/cuffee/products/dto/CoffeeResponseDto.java
@@ -8,7 +8,7 @@ import lombok.*;
 @AllArgsConstructor
 public class CoffeeResponseDto {
 
-    private int coffeeId;
+    private long coffeeId;
     private String name;
     private int price;
     private String contents;
