+++ src/main/java/com/mysite/cuffee/products/entity/Coffee.java
@@ -14,7 +14,7 @@ public class Coffee {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "coffee_id")
-    private int coffeeId;
+    private long coffeeId;
 
     @Column(nullable = false, length = 150)
     private String name;
